pyproject.toml
[project]
name = "data-engineering-revision"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dbt-core",
    "dbt-postgres",
    "dbt-mysql",
    "ipykernel",
    "jupyter",
    "dbt-snowflake",
    "dbt-bigquery",
    "dbt-spark",
    "dbt-dremio",
    "dbt-databricks",
    "mysql-connector-python>=9.4.0",
    "tqdm>=4.67.1",
    "langgraph>=0.6.5",
    "langchain-core>=0.3.74",
    "langchain-openai>=0.3.30",
    "langchain-community>=0.3.27",
    "openai>=1.99.9",
    "ipython>=8.37.0",
]
